requirements.txt
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
dbus-python==1.3.2
idna==3.4
lz4==4.3.2
requests==2.28.2
soupsieve==2.4
urllib3==1.26.15
